Fix refresh spinner functionality

+++ app/src/main/java/edu/grinnell/sandb/MainActivity.java
@@ -1,7 +1,5 @@
 package edu.grinnell.sandb;
 
-import android.app.PendingIntent;
-import android.content.BroadcastReceiver;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.AsyncTask;
@@ -17,11 +15,7 @@ import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.ViewParent;
-import android.view.animation.AnimationUtils;
-import android.widget.ImageView;
-import android.widget.Spinner;
 
 import com.crashlytics.android.Crashlytics;
 import com.flurry.android.FlurryAgent;
@@ -29,35 +23,28 @@ import com.squareup.okhttp.OkHttpClient;
 import com.squareup.okhttp.Request;
 import com.squareup.okhttp.Response;
 
-import org.json.JSONArray;
 import org.json.JSONException;
-import org.json.JSONObject;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.List;
 
 import edu.grinnell.sandb.model.Article;
-import edu.grinnell.sandb.model.Image;
 import edu.grinnell.sandb.util.BodyImageGetter;
+import edu.grinnell.sandb.util.JSONUtil;
 
 /* The main activity that the app will initialize to. This activity hosts the ArticleListFragment */
-public class MainActivity extends ActionBarActivity implements
-		ArticleListFragment.Callbacks {
+public class MainActivity extends ActionBarActivity {
 
 	private static final String TAG = "MainActivity";
 	private static final String SELECTED_TAB = "selected_tab";
 
-	private PendingIntent mSendFeedLoaded;
 	private ArticleListFragment mListFrag;
 	private View mLoading;
-	private ImageView mLoadingImage;
 	private ViewPager mPager;
 	private TabsAdapter mTabsAdapter;
 
 	private boolean mUpdateInProgress;
-	private BroadcastReceiver mUpdateReceiver;
-
-	private MainPrefs mPrefs;
 
 	private boolean mTwoPane = false;
 
@@ -67,10 +54,7 @@ public class MainActivity extends ActionBarActivity implements
 		Crashlytics.start(this);
 		setContentView(R.layout.activity_main);
 
-		mPrefs = new MainPrefs(this);
-
 		mLoading = (View) findViewById(R.id.loading);
-		mLoadingImage = (ImageView) mLoading.findViewById(R.id.spinner);
 
 		// setup action bar for tabs
 		ActionBar actionBar = getSupportActionBar();
@@ -85,21 +69,51 @@ public class MainActivity extends ActionBarActivity implements
 			actionBar.setSelectedNavigationItem(savedInstanceState
 					.getInt(SELECTED_TAB));
 		}
-
-        updateArticles();
-
 	}
 
 	@Override
 	protected void onResume() {
 		super.onResume();
-
 	}
 
+
+    @Override
+    protected void onPause() {
+        super.onPause();
+
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.activity_main, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.menu_refresh:
+                if (!mUpdateInProgress) {
+                    updateArticles();
+                }
+                break;
+            default:
+
+                break;
+        }
+
+        return false;
+    }
+
     public void updateArticles() {
-        String url = "http://www.thesandb.com/api/get_recent_posts?count=30/";
-        String[] params = {url};
-        new ArticleFetchTask().execute(params);
+        if (!mUpdateInProgress) {
+            String url = "http://www.thesandb.com/api/get_recent_posts?count=50/";
+            String[] params = {url};
+            mTabsAdapter.setRefreshing(true);
+            mUpdateInProgress = true;
+            new ArticleFetchTask().execute(params);
+        }
     }
 
     public class ArticleFetchTask extends AsyncTask<String, Void, Integer> {
@@ -125,28 +139,19 @@ public class MainActivity extends ActionBarActivity implements
                     .build();
             try {
                 Response response = client.newCall(request).execute();
-                JSONObject responseObject = new JSONObject(response.body().string());
+                List<Article> articleList = JSONUtil.parseArticleJSON(response);
 
-                Article.deleteAll(Article.class);
-                Image.deleteAll(Image.class);
-                JSONArray posts = responseObject.getJSONArray("posts");
-
-                Article newArticle;
-                for (int i = 0; i < posts.length(); ++i) {
-                    newArticle = new Article();
-                    JSONObject thisPost = posts.getJSONObject(i);
-                    newArticle.setArticleID(Integer.parseInt(thisPost.getString("id")));
-                    newArticle.setAuthor(thisPost.getJSONObject("author").getString("name"));
-                    newArticle.setBody(thisPost.getString("content"));
-                    newArticle.setDescription(thisPost.getString("excerpt"));
-                    newArticle.setTitle(thisPost.getString("title"));
-                    newArticle.setCategory(thisPost.getJSONArray("categories").getJSONObject(0).getString("title"));
-                    newArticle.setPubDate(thisPost.getString("date"));
-                    newArticle.setLink(thisPost.getString("url"));
-                    newArticle.save();
-                    BodyImageGetter.readImages(newArticle);
+                if (!articleList.isEmpty()) {
+                    Article.deleteAll(Article.class);
+                    for (Article article : articleList) {
+                        article.save();
+                        BodyImageGetter.readImages(article);
+                    }
+                    return SUCCESS;
                 }
-                return SUCCESS;
+                else {
+                    return PARSING_PROBLEMS;
+                }
             }
             catch (IOException e) {
                 Log.e(TAG, e.getMessage());
@@ -166,9 +171,9 @@ public class MainActivity extends ActionBarActivity implements
             }
             // Clear the loading bar when the articles are loaded
             if (mUpdateInProgress) {
-                    mUpdateInProgress = false;
-                    mLoading.setVisibility(View.GONE);
+                mUpdateInProgress = false;
                 }
+            mTabsAdapter.setRefreshing(false);
             mTabsAdapter.refresh();
             }
         }
@@ -179,57 +184,6 @@ public class MainActivity extends ActionBarActivity implements
             NetworkInfo n = cm.getActiveNetworkInfo();
             return (n != null) && n.isConnectedOrConnecting();
         }
-
-	@Override
-	protected void onPause() {
-		super.onPause();
-
-	}
-
-	@Override
-	public boolean onCreateOptionsMenu(Menu menu) {
-		// Inflate the menu; this adds items to the action bar if it is present.
-		getMenuInflater().inflate(R.menu.activity_main, menu);
-		return true;
-	}
-
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-		switch (item.getItemId()) {
-		case R.id.menu_refresh:
-			if (!mUpdateInProgress) {
-				// Reload the feed when the refresh button is pressed
-				mLoading.setVisibility(View.VISIBLE);
-				mLoadingImage.startAnimation(AnimationUtils.loadAnimation(this,
-						R.anim.loading));
-				mUpdateInProgress = true;
-				updateArticles();
-			}
-			break;
-		// case R.id.menu_settings:
-		// // startActivityForResult(new Intent(this, PrefActiv.class), PREFS);
-		// break;
-		default:
-
-			break;
-		} 
-
-		return false;
-	}
-
-	@Override
-	public void onItemSelected(int position) {
-		if (mTwoPane) {
-			// TODO add two pane layout
-		} else {
-			if (mSendFeedLoaded != null)
-				mSendFeedLoaded.cancel();
-		}
-	}
-
-	@Override
-	public void setListActivateState() {
-	}
 
 	// Add the catagory tabs
 	private void addTabs(ActionBar actionBar, TabsAdapter ta) {
@@ -298,40 +252,6 @@ public class MainActivity extends ActionBarActivity implements
 		public void onPageSelected(int position) {
 			mActionBar.getTabAt(position).select();
 			ViewParent root = findViewById(android.R.id.content).getParent();
-			findAndUpdateSpinner(root, position);
-		}
-
-		/**
-		 * Searches the view hierarchy excluding the content view for a possible
-		 * Spinner in the ActionBar.
-		 * 
-		 * @param root
-		 *            The parent of the content view
-		 * @param position
-		 *            The position that should be selected
-		 * @return if the spinner was found and adjusted
-		 */
-		private boolean findAndUpdateSpinner(Object root, int position) {
-			if (root instanceof android.widget.Spinner) {
-				// Found the Spinner
-				Spinner spinner = (Spinner) root;
-				spinner.setSelection(position);
-				return true;
-			} else if (root instanceof ViewGroup) {
-				ViewGroup group = (ViewGroup) root;
-				if (group.getId() != android.R.id.content) {
-					// Found a container that isn't the container holding our
-					// screen layout
-					for (int i = 0; i < group.getChildCount(); i++) {
-						if (findAndUpdateSpinner(group.getChildAt(i), position)) {
-							// Found and done searching the View tree
-							return true;
-						}
-					}
-				}
-			}
-			// Nothing found
-			return false;
 		}
 
 		@Override
@@ -368,7 +288,17 @@ public class MainActivity extends ActionBarActivity implements
 		private String getFragmentTag(int pos) {
 			return "android:switcher:" + R.id.pager + ":" + pos;
 		}
-	}
+
+        public void setRefreshing(boolean refreshing) {
+            for (int i = 0; i < getCount(); i++) {
+                Fragment f = getSupportFragmentManager().findFragmentByTag(
+                        getFragmentTag(i));
+                if (f != null)
+                    ((ArticleListFragment) f).setRefreshing(refreshing);
+            }
+        }
+    }
+
 
 	@Override
 	public void onSaveInstanceState(Bundle state) {
